jyotish-mcp/calculations/requirements.txt
# Jyotish Calculation Engine Dependencies

# Swiss Ephemeris (astronomical calculations)
pyswisseph==2.10.3.2

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code quality
black==23.12.1
ruff==0.1.8
mypy==1.7.1

# Development
ipython==8.18.1
